pyproject.toml
[project]
name = "same-image-locator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
